# Improve cluster forming speed